refactor: renomeando componente 2
